requirements.txt
openpyxl 
matplotlib 
python-telegram-bot 
